requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///D:/bld/aiohttp_1774999596133/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
annotated-doc @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_annotated-doc_1762868213/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1774359453/work
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1749017159514/work
argon2-cffi-bindings @ file:///D:/bld/argon2-cffi-bindings_1762509368570/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_async-timeout_1767290444/work
asyncpg @ file:///D:/bld/bld/rattler-build_asyncpg_1768733564/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_attrs_1773935801/work
bcrypt @ file:///D:/bld/bld/rattler-build_bcrypt_1762497737/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1776866578774/work/certifi
cffi @ file:///D:/bld/cffi_1761202893018/work
click @ file:///D:/bld/bld/rattler-build_click_1777219929/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
cryptography @ file:///D:/bld/cryptography-split_1777966038696/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_dnspython_1757292856/work
ecdsa @ file:///home/conda/feedstock_root/build_artifacts/ecdsa_1774556716531/work
email-validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1756221362582/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-core_1776987546/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-cli_1771293215/work
fastar @ file:///D:/bld/bld/rattler-build_fastar_1776177905/work
frozenlist @ file:///D:/bld/frozenlist_1752167156313/work
gmpy2 @ file:///D:/bld/gmpy2_1773244957686/work
greenlet @ file:///D:/bld/bld/rattler-build_greenlet_1777329016/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h11_1767729720/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httptools @ file:///D:/bld/httptools_1762503930761/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_idna_1776947141/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jinja2_1764517220/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1778169583724/work
MarkupSafe @ file:///D:/bld/markupsafe_1772444956616/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
motor @ file:///home/conda/feedstock_root/build_artifacts/motor_1747315692754/work
multidict @ file:///D:/bld/multidict_1771610627814/work
packaging==26.0
passlib==1.7.4
propcache @ file:///D:/bld/propcache_1744524926447/work
pyasn1 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyasn1_1773729387/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic_1778103405/work
pydantic-extra-types @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic-extra-types_1775429718/work
pydantic-settings @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic-settings_1778254612/work
pydantic_core @ file:///D:/bld/bld/rattler-build_pydantic-core_1778084255/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1774796700044/work
pymongo @ file:///D:/bld/pymongo_1776765840151/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1772388605/work
python-jose @ file:///home/conda/feedstock_root/build_artifacts/python-jose_1748530462723/work
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-multipart_1778422023/work
PyYAML @ file:///D:/bld/pyyaml_1770223251282/work
redis @ file:///home/conda/feedstock_root/build_artifacts/redis-py_1774356320455/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1775991661/work/dist
rich-toolkit @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich-toolkit_1771977622/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1744825426113/work
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1762858192781/work
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1762941432086/work
SQLAlchemy @ file:///D:/bld/bld/rattler-build_sqlalchemy_1775241408/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_starlette_1774215956/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1774492402/work
tornado @ file:///D:/bld/tornado_1774357957486/work
typer @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typer_1777583624/work
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing-inspection_1777105465/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
uvicorn @ file:///D:/bld/bld/rattler-build_uvicorn_1776948638/work
watchfiles @ file:///D:/bld/watchfiles_1760456630893/work
websockets @ file:///D:/bld/bld/rattler-build_websockets_1768087424/work/dist
yarl @ file:///D:/bld/yarl_1772409271121/work
